Implement delete function in local provider, interaction with resource model. #124

+++ app/components/Storage.php
@@ -5,6 +5,7 @@ use yii;
 use yii\base\Component;
 use yii\web\UploadedFile;
 use app\components\storageProviders\StorageProviderInterface;
+use app\models\Resource;
 
 class Storage extends Component
 {
@@ -37,10 +38,26 @@ class Storage extends Component
      * @return mixed string|bool
      */
     public function save($localPath) {
-        return  $this->provider->save($localPath);
+        $file =  $this->provider->save($localPath);
+        if ($file['path'] != null && $file['link'] != null) {
+            $resource = new Resource();
+            $resource->path = $file['path'];
+            $resource->link = $file['link'];
+            $resource->save();
+            return $resource->id;
+        } else {
+            return false;
+        }
     }
 
-    public function delete($path) {
-        return $this->provider->delete($path);
+    public function delete($id) {
+        $resource = Resource::find($id);
+        if($resource != null) {
+            $path = $resource->path;
+            $resource->delete();
+            return $this->provider->delete($path);
+        } else {
+            return false;
+        }
     }
 }
